+++ spec/brainstem/presenter_collection_spec.rb
@@ -198,6 +198,24 @@ describe Brainstem::PresenterCollection do
             end
           end
         end
+
+        describe "page_size" do
+          describe "when per_page is provided" do
+            let(:params) { { per_page: 666 } }
+
+            it "indicates the provided value" do
+              expect(result["page_size"]).to eq(666)
+            end
+          end
+
+          describe "when per_page is not provided" do
+            let(:params) { {} }
+
+            it "uses the default value" do
+              expect(result["page_size"]).to eq(@presenter_collection.default_per_page)
+            end
+          end
+        end
       end
     end
 
@@ -296,7 +314,7 @@ describe Brainstem::PresenterCollection do
     describe "the 'results' top level key" do
       it "comes back with an explicit list of the matching results" do
         structure = @presenter_collection.presenting("workspaces", :params => { :include => "tasks" }, :max_per_page => 2) { Workspace.where(:id => 1) }
-        expect(structure.keys).to match_array %w[workspaces tasks count page_number page_count results]
+        expect(structure.keys).to match_array %w[workspaces tasks count page_number page_count page_size results]
         expect(structure['results']).to eq(Workspace.where(:id => 1).limit(2).map {|w| { 'key' => 'workspaces', 'id' => w.id.to_s } })
         expect(structure['workspaces'].keys).to eq(%w[1])
       end
@@ -305,10 +323,10 @@ describe Brainstem::PresenterCollection do
     describe "includes" do
       it "reads allowed includes from the presenter" do
         result = @presenter_collection.presenting("workspaces", :params => { :include => "drop table,tasks,users" }) { Workspace.unscoped }
-        expect(result.keys).to match_array %w[count page_count page_number workspaces tasks results]
+        expect(result.keys).to match_array %w[count page_count page_number page_size workspaces tasks results]
 
         result = @presenter_collection.presenting("workspaces", :params => { :include => "foo,tasks,lead_user" }) { Workspace.unscoped }
-        expect(result.keys).to match_array %w[count page_count page_number workspaces tasks users results]
+        expect(result.keys).to match_array %w[count page_count page_number page_size workspaces tasks users results]
       end
 
       it "defaults to not include any allowed includes" do
@@ -344,7 +362,7 @@ describe Brainstem::PresenterCollection do
         result = @presenter_collection.presenting("tasks", :params => { :include => "workspace" }, :max_per_page => 2) { Task.where(:id => t.id) }
         expect(result['tasks'].keys).to eq([ t.id.to_s ])
         expect(result['workspaces']).to eq({})
-        expect(result.keys).to match_array %w[tasks workspaces count page_count page_number results]
+        expect(result.keys).to match_array %w[tasks workspaces count page_count page_number page_size results]
       end
 
       context 'when including something of the same type as the primary model' do
@@ -951,7 +969,7 @@ describe Brainstem::PresenterCollection do
       context "when there is no sort provided" do
         it "returns an empty array when there are no objects" do
           result = @presenter_collection.presenting("workspaces") { Workspace.where(:id => nil) }
-          expect(result).to eq('count' => 0, 'page_count' => 0, 'page_number' => 0, 'workspaces' => {}, 'results' => [])
+          expect(result).to eq('count' => 0, 'page_count' => 0, 'page_number' => 0, 'page_size' => 20, 'workspaces' => {}, 'results' => [])
         end
 
         it "falls back to the object's sort order when nothing is provided" do
@@ -1006,7 +1024,7 @@ describe Brainstem::PresenterCollection do
 
         result = @presenter_collection.presenting("workspaces", :params => { :order => "description:drop table" }) { Workspace.where("id is not null") }
         expect(last_direction).to eq('asc')
-        expect(result.keys).to match_array %w[count page_count page_number workspaces results]
+        expect(result.keys).to match_array %w[count page_count page_number page_size workspaces results]
 
         result = @presenter_collection.presenting("workspaces", :params => { :order => "description:;;hacker;;" }) { Workspace.where("id is not null") }
         expect(last_direction).to eq('asc')